requirements.txt
# Core dependencies
faster-whisper==1.0.3  # 4x faster than openai-whisper with multi-threading
ffmpeg-python==0.2.0
PyYAML==6.0.1

# AI/ML for GitHub Models
azure-ai-inference==1.0.0b3
openai>=1.0.0

# Google API for YouTube (legacy - being replaced with browser automation)
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP requests for API calls
requests==2.31.0

# Audio/Video processing
numpy==1.26.2

# Optional: Better performance for Whisper
# torch==2.1.1
# torchaudio==2.1.1

# Utilities
python-dotenv==1.0.0

# Local LLM support removed - using GitHub Models API only

# UI Framework
PySide6>=6.6.0

# Browser Automation
playwright>=1.40.0

# Build
pyinstaller>=6.0.0

# Advanced Viral Detection
vaderSentiment>=3.3.2  # Sentiment/emotion analysis
textblob>=0.17.1  # Natural language processing
sentence-transformers>=2.2.2  # Semantic embeddings for deduplication
